+++ projects/ngx-scrollbar/src/lib/scrollbar/scrollbar.ts
@@ -32,20 +32,25 @@ export abstract class Scrollbar implements OnInit, OnDestroy {
    * Activate scrollbar pointer events
    */
   private activatePointerEvents(): Observable<any> {
+    // Stream that emits when scrollbar thumb is dragged
     let thumbDragEvent: Observable<any> = EMPTY;
+    // Stream that emits when scrollbar track is clicked
     let trackClickEvent: Observable<any> = EMPTY;
+    // Stream that emits when scrollbar track is hovered
     let trackHoveredEvent: Observable<any> = EMPTY;
 
+    // Set the method used for the pointer events option
     if (this.cmp.pointerEventsMethod === 'viewport') {
+      // Pointer events using the viewport
       this.viewportTrackClicked = new Subject<any>();
       this.viewportThumbClicked = new Subject<any>();
 
-      // Activate viewport pointer events
-      this.cmp.viewport.activatePointerEvents();
+      // Activate the pointer events of the viewport directive
+      this.cmp.viewport.activatePointerEvents(this.cmp.viewportPropagateMouseMove, this.destroyed);
 
+      // Set streams
       thumbDragEvent = this.viewportThumbClicked;
       trackClickEvent = this.viewportTrackClicked;
-
       trackHoveredEvent = this.cmp.viewport.hovered.pipe(
         // Check if track is hovered
         map((e: any) => isWithinBounds(e, this.track.clientRect)),
@@ -56,15 +61,21 @@ export abstract class Scrollbar implements OnInit, OnDestroy {
 
       this.cmp.viewport.clicked.pipe(
         tap((e: any) => {
-          if (isWithinBounds(e, this.thumb.clientRect)) {
-            this.viewportThumbClicked.next(e);
-          } else if (isWithinBounds(e, this.track.clientRect)) {
-            this.viewportTrackClicked.next(e);
+          if (e) {
+            if (isWithinBounds(e, this.thumb.clientRect)) {
+              this.viewportThumbClicked.next(e);
+            } else if (isWithinBounds(e, this.track.clientRect)) {
+              this.cmp.setClicked(true);
+              this.viewportTrackClicked.next(e);
+            }
+          } else {
+            this.cmp.setClicked(false);
           }
         }),
         takeUntil(this.destroyed)
       ).subscribe();
     } else {
+      // Pointer events method is using 'scrollbar'
       thumbDragEvent = this.thumb.clicked;
       trackClickEvent = this.track.clicked;
       trackHoveredEvent = this.track.hovered;
@@ -87,8 +98,11 @@ export abstract class Scrollbar implements OnInit, OnDestroy {
         this.activatePointerEvents().pipe(takeUntil(this.destroyed)).subscribe();
       }
 
+      // Stream that emits when host component is updated
+      const updated = this.cmp.updated.pipe(tap(() => this.onUpdated()));
+
       // Update scrollbar thumb when viewport is scrolled and when scrollbar component is updated
-      merge(this.cmp.scrolled, this.cmp.updated).pipe(
+      merge(this.cmp.scrolled, updated).pipe(
         tap(() => this.thumb.update()),
         takeUntil(this.destroyed)
       ).subscribe();
@@ -103,11 +117,13 @@ export abstract class Scrollbar implements OnInit, OnDestroy {
     this.destroyed.complete();
 
     // Clean up viewport streams if used
-    if (this.cmp.pointerEventsMethod === 'viewport') {
+    if (this.viewportThumbClicked && this.viewportTrackClicked) {
       this.viewportTrackClicked.complete();
       this.viewportThumbClicked.complete();
     }
   }
 
   protected abstract setHovered(value: boolean): void;
+
+  protected abstract onUpdated(): void;
 }